Tweak color palette
